Clarify fraud checking $vars
It's unclear why there's a need to have an indexed array ([0] => array()) at the start of the $vars param when this function should only ever be returning a single result. Possible bug? 

isfraud is also rather misleading; implies a boolean, when it's not. Shouldn't that value simply provide a true/false after WHMCS compares the riskScore against the threshold configured?